update: post check

+++ blogs/views/staff.py
@@ -250,34 +250,35 @@ def ignore(request, pk):
 
 
 @staff_member_required
-def migrate_blog(request):    
-    subdomain = request.POST.get('subdomain')
-    email = request.POST.get('email')
-    message = ""
-    if not email or not subdomain:
-        return HttpResponse("Both email and subdomain must be provided.")
-    
-    user = User.objects.filter(email=email).first()
-    if not user:
-        return HttpResponse("User not found.")
-    message += f"Found user: {user}...<br>"
-    
-    blog = Blog.objects.filter(subdomain=subdomain).first()
-    if not blog:
-        return HttpResponse("Blog not found.")
-    message += f"Found blog: {blog.title} ({blog.useful_domain})...<br>"
-    
-    old_user = blog.user
-    message += f'Migrating blog ({blog.title}) from {old_user} to {user}...<br>'
-    blog.user = user
-    blog.save()
+def migrate_blog(request):
+    if request.method == "POST":
+        subdomain = request.POST.get('subdomain')
+        email = request.POST.get('email')
+        message = ""
+        if not email or not subdomain:
+            return HttpResponse("Both email and subdomain must be provided.")
+        
+        user = User.objects.filter(email=email).first()
+        if not user:
+            return HttpResponse("User not found.")
+        message += f"Found user: {user}...<br>"
+        
+        blog = Blog.objects.filter(subdomain=subdomain).first()
+        if not blog:
+            return HttpResponse("Blog not found.")
+        message += f"Found blog: {blog.title} ({blog.useful_domain})...<br>"
+        
+        old_user = blog.user
+        message += f'Migrating blog ({blog.title}) from {old_user} to {user}...<br>'
+        blog.user = user
+        blog.save()
 
-    if old_user.blogs.count() == 0:
-        message += f'User {old_user} has no more blogs and will be deleted...<br>'
-        old_user.delete()
-        message += 'Deleted...\n'
-    
-    return HttpResponse(message)
+        if old_user.blogs.count() == 0:
+            message += f'User {old_user} has no more blogs and will be deleted...<br>'
+            old_user.delete()
+            message += 'Deleted...\n'
+        
+        return HttpResponse(message)
